add: address checker component
