+++ components/transfer/Transfer.vue
@@ -111,7 +111,7 @@
                 },
               ]"
               placeholder="Enter wallet address"
-              with-address-checker />
+              empty-on-error />
             <NeoInput
               v-if="displayUnit === 'token'"
               v-model="destinationAddress.token"
@@ -132,6 +132,16 @@
               icon-right-class="search"
               class="is-flex-1"
               @input="onUsdFieldChange(destinationAddress)" />
+          </div>
+          <div>
+            <AddressChecker
+              :address="destinationAddress.address"
+              @check="
+                (isValid) => handleAddressCheck(destinationAddress, isValid)
+              "
+              @change="
+                (address) => handleAddressChange(destinationAddress, address)
+              " />
           </div>
         </div>
       </div>
@@ -285,6 +295,7 @@ import AddressInput from '@/components/shared/AddressInput.vue'
 import TransactionLoader from '@/components/shared/TransactionLoader.vue'
 import { KODADOT_DAO } from '@/utils/support'
 import { toDefaultAddress } from '@/utils/account'
+import AddressChecker from '@/components/shared/AddressChecker.vue'
 
 const Money = defineAsyncComponent(
   () => import('@/components/shared/format/Money.vue')
@@ -317,6 +328,7 @@ export type TargetAddress = {
   address: string
   usd?: number | string
   token?: number | string
+  isInvalid?: boolean
 }
 
 const isMobile = computed(() => useWindowSize().width.value <= 1024)
@@ -337,7 +349,9 @@ const tokenTabs = ref<TransferTokenTab[]>([])
 const targetAddresses = ref<TargetAddress[]>([{ address: '' }])
 
 const hasValidTarget = computed(() =>
-  targetAddresses.value.some((item) => isAddress(item.address) && item.token)
+  targetAddresses.value.some(
+    (item) => isAddress(item.address) && !item.isInvalid && item.token
+  )
 )
 
 const getDisplayUnitBasedValues = (
@@ -515,6 +529,18 @@ const onUsdFieldChange = (target: TargetAddress) => {
   if (sendSameAmount.value) {
     unifyAddressAmount(target)
   }
+}
+
+const handleAddressCheck = (target: TargetAddress, isValid: boolean) => {
+  target.isInvalid = !isValid
+
+  targetAddresses.value = [...targetAddresses.value]
+}
+
+const handleAddressChange = (target: TargetAddress, newAddress: string) => {
+  target.address = newAddress
+
+  targetAddresses.value = [...targetAddresses.value]
 }
 
 const unifyAddressAmount = (target: TargetAddress) => {
